+++ src/main/java/com/leafpage/controller/book/BookYController.java
@@ -1,7 +1,7 @@
 package com.leafpage.controller.book;
 
 import com.leafpage.controller.Controller;
-import com.leafpage.dao.BookDAO;
+import com.leafpage.gwakdao.BookDAO;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
