+++ src/main/java/com/leafpage/controller/user/MypageInfoController.java
@@ -1,14 +1,13 @@
 package com.leafpage.controller.user;
 
 import com.leafpage.controller.Controller;
-import com.leafpage.dao.BookDAO;
-import com.leafpage.dto.MypageBooksDTO;
-import com.leafpage.dto.MypageReturnedBooksDTO;
+import com.leafpage.gwakdao.BookDAO;
+import com.leafpage.gwakdto.MypageBooksDTO;
+import com.leafpage.gwakdto.MypageReturnedBooksDTO;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
-import java.awt.print.Book;
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
