--- src/main/java/com/leafpage/dto/MypageBooksDTO.java
@@ -1,21 +1,0 @@
-package com.leafpage.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class MypageBooksDTO {
-
-    int all_rentals;
-    String book_name;
-    String author_name;
-    String scheduled_return_date;
-    String rental_date;
-    String rental_no;
-    int pcY;
-    int modalWidth;
-
-}
