--- src/main/java/com/leafpage/dto/UserDTO.java
@@ -1,4 +1,0 @@
-package com.leafpage.dto;
-
-public class UserDTO {
-}
