+++ src/main/java/com/leafpage/gwakdto/MypageBooksDTO.java
@@ -1,0 +1,21 @@
+package com.leafpage.gwakdto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class MypageBooksDTO {
+
+    int all_rentals;
+    String book_name;
+    String author_name;
+    String scheduled_return_date;
+    String rental_date;
+    String rental_no;
+    int pcY;
+    int modalWidth;
+
+}
