+++ src/main/java/com/leafpage/gwakdto/UserDTO.java
@@ -1,0 +1,4 @@
+package com.leafpage.gwakdto;
+
+public class UserDTO {
+}
